Added movesearching in parallel directions, bugfixes.
We can now build a word parallel to other words.  This is done
by creating two followstructs for every candidate:  one tries
moving 'away' from the board, but another tries to move alongside
it.  This involved functionality to only consider moves after
ensuring that adding a tile to a location works in every direction.

There's still likely a bug for when you need to connect two islands.
This case will be investigated in the future, with wildcards.