Fixed a bug generating incorrect moves.
We would flip a followstruct if possible without first checking whether or
not it would lead to an impermissible move, since we would do the check whether
or not to flip after a call to followstruct:next without checks to whether we were
connecting islands or not.  This is demonstrated in sigma_perpendicular_test.

While I was poking around, I cleaned up the code a little.  The relevant changes however
are to get_move_helper.