Fixed a bug with bad words forming.
We were adding to the "completed word" list after adding a tile, but without
checking that tile's surroundings -in the direction it was heading in.-  So
despite passing the test for being perpendicular, if we connected two islands
it wouldn't matter if those islands connected!  The new test case demonstrates
this case.