Added required Thrift binaries, more build.
Added the required Thrift binaries for the generated Thrift files to
work.  Updated the build script to compile the generated Thrift files.
Updated README to state the requirements in a separate section, and
TODO to document the steps I'm taking with this refactor.

Note that the Thrift binaries come straight from the main Thrift package.
The steps for building them come from the tutorial.