Added Ruby Thrift files, included paths in source.
We need the Ruby lib/ files of Thrift to use the generated files,
so they're included in lib/, just like the Erlang ones.  We also
include all the files in the require path of conversationalist.rb,
who is most like to use them in the current client.

By the way, the current client is a pile of dung.  It may not be
worth salvaging.