Add implementation for suggestions.  Need test.
Incidentally, we ran into a design problem with how to handle the
search function:  We want the server to create it once and query
it forever, but there's no way to alert Thrift to the need of this
data structure to loop with continuously as in the previous design.

We solve this by storing it persistently in a named ETS table.
Originally I just wanted to make a process that listens and serves
the function on-demand, but then we run into the same problem with
the Pid of that process :-/.