Client talks to server.
We can also draw a board in a rudimentary fashion, though
reading from the server isn't happening how I'd like.  The new
libthrift.swc in this commit contains a basic wrapper to socket
which I hope to use as a basic client for Thrift, since out-of-the-box
it only supports and HTTP client.

The rebar.config allows for debugging symbols in the build, which
you need for Dialyzer analysis.