Steps toward 1-button builds of whole system.
Currently we have subprojects, each with their own Makefile.  I'd like
to configure things so I can push one button and build, test
everything.  This would also mean build output is in one convenient
place.

This means writing a fair bit of glue between the normal build commands
of the subprojects, and (more laboriously) making sure each
subproject's output can be structured in a way that you can simply
navigate to 'build' and run the program, or a test suite.

We start by taking steps toward the OTP release model for the server,
and making the top-level Makefile build/clean the subprojects. While
I'm still on the fence as to whether or not Make is the best system
for this, it's the one where I use the least immediate effort :-/
With more time, I can investigate alternatives.