Common Test integration, part 1.
Since we're testing the top-level server rather than babby components,
we'll need something more robust than EUnit. I'm moving towards
Common Test, but rebar support for it is lame (you can't even specify
the directory to search for...).

This change has the first integrations, where we run the tests.  They
fail, naturally, and the build process to get them to work is dependent
on building a full release, which we don't check for ahead of time
gracefully either.  We're about to outgrow makefiles as well...

This leaves a lot of open questions: should code coverage be provided
by eunit + rebar (it works!) or should I migrate all Eunit tests to
Common Test and use that code coverage?

Interesting, frightening, exciting times ahead!