Build utilities, play_move tests.
Testing was become a pain in the ass because we were starting and
stopping the server each time.  This took about 10-15 seconds because
the server had to load the GADDAG objects.  This made testing more
laborious.

Tests also weren't working because the capacity to truly support
multiple dictionaries wasn't 'baked in': we only worked with the
assumption that the TWL dictionary/gaddag was in use.

This change includes a new server_util, that will only stop or start
the server as necessary. It also adds support for multiple
dictionaries, though my dev machines don't have the memory to handle
the use of all of them yet :-/

Also, tests for play_move on all the games have been added.  WWF tests
are disabled, until I can work on GADDAG minimization or get more
memory.