Final commit before branching off for memory fixes.
The memory usage of this application is bezerk: my abstractions
are broken.  A bugfix in scrabblecheat_main (commented out, the
line is gamestate:verify(WithMove)) is now causing the program
to run out of memory.

Memory issues also plague the use of other gaddags, games longer
than a few moves crash. Time to see what's going on...