First phase of memory refactor, lots o' improvements
The major culprit of the memory loss (to my estimation) was the use of
ETS tables - when copying a large binary from an ETS table, it's
copied into process space.

This commit is the first part of my planned fix: a separate module
to handle the GADDAGs. It's not passing tests at the moment, but I
want to update my branches, dammit.