Add bingad and erlang binary reader code.
The server runs, but is useless. Need to look at the hastily-removed
gaddag_looper and put that code in the right place, with the new
modules.

Also, fucking Lion broke my Ruby gems, so the client that I had working
is bunk atm. Since upgrading to Lion, it's been nothing but one hairy
yak after another. Lovingly: fuck yourselves, Apple.