Bin_trie supports multiple games. With Douchery.
It's something of a hack. Or, if you're optimistic, the benefits of
abstraction! I used to be passing around binary bits that reference
the giant binary in the middle.  Now they're a pair, {Binary, Name},
since the giant binary needs to know whether we're looking into
twl06, sowpods, or zynga dictionaries at every operation.

The benefits of abstraction! Since this only happens at the bin_trie
module, changing the data was EZ. In your face, OOP.