Add test case, modify Makefile to support. Fails.
The test case just makes an HTTP client, calls our HTTP server,
and ensures the data is valid. The HTTP client dies because it
'cannot refill buffer,' so it's soon time to dive into the
implementation of THttpClient.

In the meantime, the build system is jankier than normal. Dev life.