'Fixed' HTTP test, it works!
Hacked up another HTTP client using the Thrift Ruby bindings rather than
the C++ ones -- for some reason this worked with the C++ HTTP server.
Devised a very simple HTTP test on the 'new_game' function, and our
server passes.

Since we're testing the Thrift Basic, non-HTTP server in the Erlang
with the Common Test anyways, it probably makes more sense to leverage
that code and have it run all the same tests, but with an HTTP Thrift
client instead. However, it's comforting to have this commit in here
anyways, since it represents a state when it all works :-p