Bugfix: Added another layer of security for bad boards.
When trying to handle a test case for my sister's WWF game, I input the board wrong
but the board:verify code didn't catch it, causing the movesearch to puke and die.
We add another layer to throw the exception if the movesearch is unable to produce
the GADDAGs necessary to continue due to bad boards.

In the future -- a) board:verify should be more bulletproof, and b) we should allow
movesearches on bad boards anyways, since players can bluff.