Fixed bad Zynga dictionary.
The GADDAG produced from the Zynga dictionary was full of bad data,
and since we're operating on raw binaries, there wasn't a typechecker to save
use from certain doom -- the only prescription being a memory hole until heap
overflow. The 'bingad' app is supposed to convert to upper case, but I have a
feeling my homebrew solution corrupted the resulting .dict file.

Quick fix, since I want to be working on clients now, is to simply upcase the
Zynga dict and run it through like the rest of them. It worked! I was able to
re-enable the WWF test case in the CT server suite, and ran Annalisa's data
through it.

CLIENT TIME!