Better install, license/doc changes (#13)
Even the "restoration" branch didn't quite finish the work: the project
docs were stale, I'd like my software to be Free, and trying to get this
running on a new machine pointed at a ton of little problems. This
commit adds Bundler support to the Ruby client and kills a ton of
old files, adds a Makefile for the server so we can handle the C/Bingad
interaction better, and otherwise makes it more plug-and-play.